refactoring to consolidate configuration code
