+++ src/Ocelot/Configuration/Creator/IInternalConfigurationCreator.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+using Ocelot.Configuration.File;
+using Ocelot.Responses;
+
+namespace Ocelot.Configuration.Creator
+{
+    public interface IInternalConfigurationCreator
+    {
+        Task<Response<IInternalConfiguration>> Create(FileConfiguration fileConfiguration);
+    }
+}
