--- src/Ocelot/Configuration/IOcelotConfiguration.cs
@@ -1,12 +1,0 @@
-using System.Collections.Generic;
-
-namespace Ocelot.Configuration
-{
-    public interface IOcelotConfiguration
-    {
-        List<ReRoute> ReRoutes { get; }
-        string AdministrationPath {get;}
-        ServiceProviderConfiguration ServiceProviderConfiguration {get;}
-        string RequestId {get;}
-    }
-}
