--- src/Ocelot/Configuration/Provider/IOcelotConfigurationProvider.cs
@@ -1,9 +1,0 @@
-﻿using Ocelot.Responses;
-
-namespace Ocelot.Configuration.Provider
-{
-    public interface IOcelotConfigurationProvider
-    {
-        Response<IOcelotConfiguration> Get();
-    }
-}
