+++ src/Ocelot/Configuration/Repository/IInternalConfigurationRepository.cs
@@ -1,0 +1,10 @@
+﻿using Ocelot.Responses;
+
+namespace Ocelot.Configuration.Repository
+{
+    public interface IInternalConfigurationRepository
+    {
+        Response<IInternalConfiguration> Get();
+        Response AddOrReplace(IInternalConfiguration internalConfiguration);
+    }
+}
