--- src/Ocelot/Configuration/Repository/IOcelotConfigurationRepository.cs
@@ -1,10 +1,0 @@
-﻿using Ocelot.Responses;
-
-namespace Ocelot.Configuration.Repository
-{
-    public interface IOcelotConfigurationRepository
-    {
-        Response<IOcelotConfiguration> Get();
-        Response AddOrReplace(IOcelotConfiguration ocelotConfiguration);
-    }
-}
