+++ test/Ocelot.UnitTests/DownstreamRouteFinder/DownstreamRouteFinderMiddlewareTests.cs
@@ -1,7 +1,4 @@
-﻿using Ocelot.Middleware;
-using Ocelot.Middleware.Multiplexer;
-
-namespace Ocelot.UnitTests.DownstreamRouteFinder
+﻿namespace Ocelot.UnitTests.DownstreamRouteFinder
 {
     using System.Collections.Generic;
     using System.Threading.Tasks;
@@ -9,7 +6,6 @@ namespace Ocelot.UnitTests.DownstreamRouteFinder
     using Moq;
     using Ocelot.Configuration;
     using Ocelot.Configuration.Builder;
-    using Ocelot.Configuration.Provider;
     using Ocelot.DownstreamRouteFinder;
     using Ocelot.DownstreamRouteFinder.Finder;
     using Ocelot.DownstreamRouteFinder.Middleware;
@@ -19,23 +15,26 @@ namespace Ocelot.UnitTests.DownstreamRouteFinder
     using Shouldly;
     using TestStack.BDDfy;
     using Xunit;
+    using Ocelot.Configuration.Repository;
+    using Ocelot.Middleware;
+    using Ocelot.Middleware.Multiplexer;
 
     public class DownstreamRouteFinderMiddlewareTests
     {
         private readonly Mock<IDownstreamRouteFinder> _finder;
-        private readonly Mock<IOcelotConfigurationProvider> _provider;
+        private readonly Mock<IInternalConfigurationRepository> _repo;
         private Response<DownstreamRoute> _downstreamRoute;
-        private IOcelotConfiguration _config;
+        private IInternalConfiguration _config;
         private Mock<IOcelotLoggerFactory> _loggerFactory;
         private Mock<IOcelotLogger> _logger;
-        private DownstreamRouteFinderMiddleware _middleware;
-        private DownstreamContext _downstreamContext;
+        private readonly DownstreamRouteFinderMiddleware _middleware;
+        private readonly DownstreamContext _downstreamContext;
         private OcelotRequestDelegate _next;
         private readonly Mock<IMultiplexer> _multiplexer;
 
         public DownstreamRouteFinderMiddlewareTests()
         {
-            _provider = new Mock<IOcelotConfigurationProvider>();
+            _repo = new Mock<IInternalConfigurationRepository>();
             _finder = new Mock<IDownstreamRouteFinder>();
             _downstreamContext = new DownstreamContext(new DefaultHttpContext());
             _loggerFactory = new Mock<IOcelotLoggerFactory>();
@@ -43,13 +42,13 @@ namespace Ocelot.UnitTests.DownstreamRouteFinder
             _loggerFactory.Setup(x => x.CreateLogger<DownstreamRouteFinderMiddleware>()).Returns(_logger.Object);
             _next = context => Task.CompletedTask;
             _multiplexer = new Mock<IMultiplexer>();
-            _middleware = new DownstreamRouteFinderMiddleware(_next, _loggerFactory.Object, _finder.Object, _provider.Object, _multiplexer.Object);
+            _middleware = new DownstreamRouteFinderMiddleware(_next, _loggerFactory.Object, _finder.Object, _repo.Object, _multiplexer.Object);
         }
 
         [Fact]
         public void should_call_scoped_data_repository_correctly()
         {
-            var config = new OcelotConfiguration(null, null, new ServiceProviderConfigurationBuilder().Build(), "");
+            var config = new InternalConfiguration(null, null, new ServiceProviderConfigurationBuilder().Build(), "");
 
             var downstreamReRoute = new DownstreamReRouteBuilder()
                 .WithDownstreamPathTemplate("any old string")
@@ -74,19 +73,19 @@ namespace Ocelot.UnitTests.DownstreamRouteFinder
             _middleware.Invoke(_downstreamContext).GetAwaiter().GetType();
         }
 
-        private void GivenTheFollowingConfig(IOcelotConfiguration config)
+        private void GivenTheFollowingConfig(IInternalConfiguration config)
         {
             _config = config;
-            _provider
+            _repo
                 .Setup(x => x.Get())
-                .Returns(new OkResponse<IOcelotConfiguration>(_config));
+                .Returns(new OkResponse<IInternalConfiguration>(_config));
         }
 
         private void GivenTheDownStreamRouteFinderReturns(DownstreamRoute downstreamRoute)
         {
             _downstreamRoute = new OkResponse<DownstreamRoute>(downstreamRoute);
             _finder
-                .Setup(x => x.FindDownstreamRoute(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<IOcelotConfiguration>(), It.IsAny<string>()))
+                .Setup(x => x.FindDownstreamRoute(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<IInternalConfiguration>(), It.IsAny<string>()))
                 .Returns(_downstreamRoute);
         }
